requirements.txt
PyYAML~=6.0
setuptools~=59.6.0
pandas~=1.1.5
Flask~=2.0.3
flask_restful
testing.postgresql
pytest
pg8000~=1.26.0
confluent_kafka
aniso8601
confluent-kafka
psycopg2-binary
pytest-postgresql
ray~=1.12.0
pre-commit